Implement Handler
I created windowsauthtokenhandler_windows.go and implements the HTTP
handler that will extract the token from the X-IIS-WindowsAuthToken
header for the request.

I created windowsauthtokenhandler_notwindows.go to implement a
passthrough handler for consuming applications when compiled for
non-Windows platforms.

I updated README.md to change the naming of the handler from
WindowsAuthTokenHandler to just Handler.

I created doc.go and added documentation for the package.